backend/requirements-minimal.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psutil==5.9.6
fastapi-cache2==0.2.1
bcrypt==4.0.1
slowapi==0.1.9
email-validator==2.1.0
httpx==0.25.2